Move store credit methods out of order and into module
